preparation/requirements.txt
airtable-python-wrapper==0.13.0
arrow==0.15.5
astroid==2.3.3
beautifulsoup4==4.8.2
bibtexparser==1.1.0
blis==0.4.1
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
crossrefapi==1.4.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
cytoolz==0.10.1
decorator==4.4.2
elsapy==0.5.0
future==0.18.2
google-search-results==1.7.1
idna==2.9
importlib-metadata==1.5.2
isort==4.3.21
jellyfish==0.7.2
joblib==0.14.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
murmurhash==1.0.2
networkx==2.4
nltk==3.4.5
numpy==1.18.2
pandas==1.0.3
plac==1.1.3
preshed==3.0.2
pycparser==2.20
pyemd==0.5.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
Pyphen==0.9.5
python-dateutil==2.8.1
python-doi==0.2.0
pytz==2019.3
requests==2.23.0
rispy==0.5.0
scholarly==0.2.5
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
toolz==0.10.0
tqdm==4.43.0
typed-ast==1.4.1
urllib3==1.25.8
wasabi==0.6.0
wrapt==1.11.2
xlrd==1.2.0
zipp==3.1.0
